rename: tabs 정리
